+++ README.md
@@ -1,4 +1,4 @@
-## 🤖 SwiftSage (v2): 
+# 🤖 SwiftSage (v2): 
 
 > [!IMPORTANT]
 > The code of SwiftSage v1 (for the experiments in NeurIPS 2023) is archived in the [`science_world`](https://github.com/SwiftSage/SwiftSage/tree/science_world) branch.
@@ -6,8 +6,22 @@
 
 <!-- Github Readme Important Callout box note -->
 
+## Installation
 
 ```bash
-# generate requirements.txt
-pip freeze > requirements.txt
+pip install -e .
+```
+
+## Usage
+
+Commandline:
+
+```bash
+swiftsage -p "which one is larger? 9.11 or 9.8?"
+```
+
+Gradio:
+
+```bash
+python -m swiftsage.app
 ```
