--- requirements.txt
@@ -1,15 +1,0 @@
-gradio==4.19.2
-numpy==1.26.4
-openai==1.46.0
-fuzzywuzzy==0.18.0
-scikit-learn==1.4.2
-dirtyjson==1.0.8
-hjson==3.1.0
-datasets==2.17.1
-sympy==1.12.1
-latex2sympy2==1.9.1
-Pebble==5.0.7
-tqdm==4.66.5
-word2number==1.1
-regex==2024.4.16
-multiprocess==0.70.16
