+++ swiftsage/utils/utils.py
@@ -3,7 +3,6 @@ import json
 import logging
 import os
 import re
-from abc import ABC, abstractmethod
 
 import dirtyjson
 import hjson
@@ -24,14 +23,6 @@ api_configs = {
     # You can add more API configurations here for other providers
 }
 
-class Agent(ABC):
-    def __init__(self, prompt_template, llm_client):
-        self.prompt_template = prompt_template
-        self.llm_client = llm_client
-
-    @abstractmethod
-    def generate_response(self, prompt):
-        pass
 
 
 def setup_logging():
